third_party/dawn/third_party/catapult/third_party/vinn/third_party/parse5/test/data/tree_construction/main-element.dat

#data
<!doctype html><p>foo<main>bar<p>baz
#errors
36: End of file seen and there were open elements.
27: Unclosed element “main”.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|     <main>
|       "bar"
|       <p>
|         "baz"

#data
<!doctype html><main><p>foo</main>bar
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <main>
|       <p>
|         "foo"
|     "bar"
